#180 further suggestions for improvement - begin restructuring command
