+++ Assets/SEE/Controls/Actions/DeleteAction.cs
@@ -163,7 +163,7 @@ namespace SEE.Controls.Actions
                 if (selectedObject.CompareTag(Tags.Edge))
                 {
                     List<GameObject> edge = new List<GameObject>() { selectedObject };
-                    actionHistory.SaveObjectForDeleteUndo(edge, new List<Vector3>());
+                    SaveObjectForDeleteUndo(edge, new List<Vector3>());
                 }
                 else if (selectedObject.CompareTag(Tags.Node))
                 {
@@ -183,29 +183,43 @@ namespace SEE.Controls.Actions
         /// </summary>
         public override void Undo()
         {
-            DeleteAction deleteAction = (DeleteAction)actionHistory.ActionHistoryList.Last();
-            Graph graph = deleteAction.Graph;
-            foreach (GameObject node in deleteAction.DeletedNodes)
+            foreach (GameObject node in DeletedNodes)
             {
                 if (node.TryGetComponentOrLog(out NodeRef nodeRef))
                 {
-                    if (!graph.Contains(nodeRef.Value))
+                    if (!Graph.Contains(nodeRef.Value))
                     {
-                        graph.AddNode(nodeRef.Value);
+                        Graph.AddNode(nodeRef.Value);
                     }
                 }
             }
 
-            foreach (GameObject edge in deleteAction.DeletedEdges)
+            foreach (GameObject edge in DeletedEdges)
             {
                 if (edge.TryGetComponentOrLog(out EdgeRef edgeReference))
                 {
-                    graph.AddEdge(edgeReference.edge);
+                    Graph.AddEdge(edgeReference.edge);
                     edge.SetVisibility(true, false);
                 }
             }
 
             actionHistory.ActionHistoryList.RemoveLast();
+        }
+
+        /// <summary>
+        /// Redoes this DeleteAction
+        /// </summary>
+        public override void Redo()
+        {
+            throw new NotImplementedException();
+        }
+
+        /// <summary>
+        /// Executes this DeleteAction
+        /// </summary>
+        public override void Execute()
+        {
+            throw new NotImplementedException();
         }
 
         /// <summary>
@@ -225,7 +239,7 @@ namespace SEE.Controls.Actions
                     Portal.SetInfinitePortal(deletedNode);
                 }
             }
-            actionHistory.SaveObjectForDeleteUndo(deletedNodes, oldPositions);
+            SaveObjectForDeleteUndo(deletedNodes, oldPositions);
             foreach (GameObject deletedNode in deletedNodes)
             {
                 Tweens.Move(deletedNode, new Vector3(garbageCan.transform.position.x, garbageCan.transform.position.y + 1.4f, garbageCan.transform.position.z), TimeForAnimation);
@@ -252,9 +266,6 @@ namespace SEE.Controls.Actions
         /// <returns>the waiting time between moving deleted nodes from the garbage-can and then to the city</returns>
         private IEnumerator RemoveNodeFromGarbage(List<GameObject> deletedNodes)
         {
-            DeleteAction deleteAction = (DeleteAction)actionHistory.ActionHistoryList.Last();
-            List<Vector3> oldPositionsOfDeletedObjects = deleteAction.OldPositions;
-
             // In case that the deleted object is a single edge - waiting-time is unintended
             if (deletedNodes.Count != 0)
             {
@@ -267,13 +278,79 @@ namespace SEE.Controls.Actions
 
                 for (int i = 0; i < deletedNodes.Count; i++)
                 {
-                        Tweens.Move(deletedNodes[i], oldPositionsOfDeletedObjects[i], TimeForAnimation);
+                        Tweens.Move(deletedNodes[i], OldPositions[i], TimeForAnimation);
                 }
 
                 yield return new WaitForSeconds(TimeToWait);
             }
             Undo();
             InteractableObject.UnselectAll(true);
+        }
+
+        /// <summary>
+        /// Saves the deleted nodes and/or edges for the possibility of an undo. 
+        /// Removes the gameObjects from the graph.
+        /// Precondition: <see cref="deletedNodes""/> != null.
+        /// </summary>
+        /// <param name="deletedNodes">all deleted nodes of the last operation</param>
+        /// <param name="oldPositionsOfDeletedNodes">all old positions of the deleted nodes of the last operation</param>
+        public void SaveObjectForDeleteUndo(List<GameObject> deletedNodes, List<Vector3> oldPositionsOfDeletedNodes)
+        {
+            SEECity city = SceneQueries.GetCodeCity(deletedNodes[0].transform)?.gameObject.GetComponent<SEECity>();
+            Graph graph = city.LoadedGraph;
+            List<GameObject> nodesAndAscendingEdges = new List<GameObject>();
+            List<GameObject> edgesToHide = new List<GameObject>();
+
+            foreach (GameObject actionHistoryObject in deletedNodes)
+            {
+                if (actionHistoryObject.TryGetComponent(out NodeRef nodeRef))
+                {
+                    HashSet<string> edgeIDs = Destroyer.GetEdgeIds(nodeRef);
+                    foreach (GameObject edge in GameObject.FindGameObjectsWithTag(Tags.Edge))
+                    {
+                        if (edge.activeInHierarchy && edgeIDs.Contains(edge.name))
+                        {
+                            edge.SetVisibility(false, true);
+
+                            if (!nodesAndAscendingEdges.Contains(edge))
+                            {
+                                edgesToHide.Add(edge);
+                            }
+                            edge.TryGetComponent(out EdgeRef edgeRef);
+                            graph.RemoveEdge(edgeRef.edge);
+                        }
+                    }
+
+                    nodesAndAscendingEdges.Add(actionHistoryObject);
+                }
+            }
+
+            List<GameObject> deletedNodesReverse = deletedNodes;
+            // For deletion bottom-up
+            deletedNodesReverse.Reverse();
+
+            foreach (GameObject deletedNode in deletedNodesReverse)
+            {
+                if (deletedNode.CompareTag(Tags.Node))
+                {
+                    deletedNode.TryGetComponent(out NodeRef nodeRef);
+                    if (graph.Contains(nodeRef.Value))
+                    {
+                        graph.RemoveNode(nodeRef.Value);
+                    }
+                }
+                if (deletedNode.CompareTag(Tags.Edge))
+                {
+                    deletedNode.SetVisibility(false, true);
+                    edgesToHide.Add(deletedNode);
+                    deletedNode.TryGetComponent(out EdgeRef edgeRef);
+                    graph.RemoveEdge(edgeRef.edge);
+                }
+            }
+
+            oldPositionsOfDeletedNodes.Reverse();
+            nodesAndAscendingEdges.Reverse();
+            actionHistory.ActionHistoryList.AddLast(new DeleteAction(nodesAndAscendingEdges, oldPositionsOfDeletedNodes, edgesToHide, graph));
         }
