+++ Assets/SEE/Controls/Actions/EditNodeAction.cs
@@ -119,6 +119,22 @@ namespace SEE.Controls.Actions
         {
             throw new NotImplementedException();
         }
+
+        /// <summary>
+        /// Redoes this DeleteAction
+        /// </summary>
+        public override void Redo()
+        {
+            throw new NotImplementedException();
+        }
+
+        /// <summary>
+        /// Executes this DeleteAction
+        /// </summary>
+        public override void Execute()
+        {
+            throw new NotImplementedException();
+        }
     }
 }
